move artifact generator to new module
