+++ legend-engine-config/legend-engine-server/pom.xml
@@ -335,6 +335,10 @@
         </dependency>
         <dependency>
             <groupId>org.finos.legend.engine</groupId>
+            <artifactId>legend-engine-xt-snowflakeApp-generator</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.finos.legend.engine</groupId>
             <artifactId>legend-engine-xt-hostedService-compiler</artifactId>
             <scope>runtime</scope>
         </dependency>
