--- legend-engine-xts-snowflakeApp/legend-engine-xt-snowflakeApp-api/src/main/resources/META-INF/services/org.finos.legend.engine.language.pure.dsl.generation.extension.ArtifactGenerationExtension
@@ -1,1 +1,0 @@
-org.finos.legend.engine.language.snowflakeApp.deployment.SnowflakeAppArtifactGenerationExtension
