+++ legend-engine-xts-snowflakeApp/legend-engine-xt-snowflakeApp-generator/src/main/resources/META-INF/services/org.finos.legend.engine.language.pure.dsl.generation.extension.ArtifactGenerationExtension
@@ -1,0 +1,1 @@
+org.finos.legend.engine.language.snowflakeApp.generator.SnowflakeAppArtifactGenerationExtension
